Pull 3d globe implementation into this repo
Associated with: https://trello.com/c/JQAnjuGp/12-load-data-of-the-cooperatives-in-the-global-map

This commit removes the iframe implementation of the 3d globe, opting
for in-repo JS code. The container is not sized correctly for some
reason. And the mouse events are captured by the 3d globe, so we may
want to remove that? Currently only countries are labeled in the globe,
as defined in 3dglobe/assets/data/countries.js and the connections
between these countries are defined in
3dglobe/assets/data/connections.js

Further styling of the globe, dots, markers, lines, shaders, etc can be
done in 3dglobe/scripts/*